refactor: package 명 오타 수정
