+++ src/main/java/nextstep/subway/application/StationService.java
@@ -1,7 +1,7 @@
-package nextstep.subway.applicaion;
+package nextstep.subway.application;
 
-import nextstep.subway.applicaion.dto.StationRequest;
-import nextstep.subway.applicaion.dto.StationResponse;
+import nextstep.subway.application.dto.StationRequest;
+import nextstep.subway.application.dto.StationResponse;
 import nextstep.subway.domain.Station;
 import nextstep.subway.domain.StationRepository;
 import nextstep.subway.domain.exception.StationException;
