+++ src/main/java/nextstep/subway/application/dto/StationResponse.java
@@ -1,4 +1,4 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.application.dto;
 
 import nextstep.subway.domain.Station;
